+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/edit.tsx
@@ -8,24 +8,28 @@ import {
 	InnerBlocks,
 } from '@wordpress/block-editor';
 import { PanelBody, ColorPicker } from '@wordpress/components';
+import { type BlockEditProps } from '@wordpress/blocks';
+
+export type Attributes = {
+	color: string;
+	storeOnly: boolean;
+};
+
+export type EditProps = BlockEditProps< Attributes >;
 
 /**
  * Internal dependencies
  */
-import { generateStyles } from './styles';
 
-export default function Edit( { attributes, setAttributes } ) {
+export default function Edit( { attributes, setAttributes }: EditProps ) {
 	const { color, storeOnly } = attributes;
 	const blockProps = { ...useBlockProps() };
 
 	if ( storeOnly ) {
 		return (
-			<>
-				<div { ...blockProps }>
-					<InnerBlocks />
-				</div>
-				<style>{ `.woocommerce-breadcrumb {display: none;}` }</style>
-			</>
+			<div { ...blockProps }>
+				<InnerBlocks />
+			</div>
 		);
 	}
 
@@ -35,6 +39,7 @@ export default function Edit( { attributes, setAttributes } ) {
 				<PanelBody title={ __( 'Settings', 'woocommerce' ) }>
 					<ColorPicker
 						color={ color }
+						// @ts-expect-error type is not defined in the library
 						onChange={ ( newColor: string ) =>
 							setAttributes( { color: newColor } )
 						}
@@ -45,8 +50,8 @@ export default function Edit( { attributes, setAttributes } ) {
 			</InspectorControls>
 			<div { ...blockProps }>
 				<InnerBlocks />
+				<style>{ `:root{--woocommerce-coming-soon-color: ${ color } }` }</style>
 			</div>
-			<style>{ generateStyles( color ) }</style>
 		</>
 	);
 }
