+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/entire-site.scss
@@ -1,0 +1,172 @@
+body:has(.woocommerce-coming-soon-entire-site) {
+	margin: 0;
+	background-color: var(--woocommerce-coming-soon-color);
+	font-family: Inter, sans-serif;
+	min-width: 320px;
+	--wp--preset--color--contrast: #111;
+	--wp--style--global--wide-size: 1280px;
+
+	/* Reset */
+	h1,
+	p,
+	a {
+		margin: 0;
+		padding: 0;
+		border: 0;
+		vertical-align: baseline;
+	}
+
+	ol,
+	ul {
+		list-style: none;
+	}
+
+	a {
+		text-decoration: none;
+	}
+
+	.is-layout-constrained > .alignwide {
+		margin: 0 auto;
+	}
+
+	.woocommerce-coming-soon-banner.alignwide {
+		max-width: 820px;
+	}
+}
+
+.block-editor-block-preview__content-iframe body {
+	/* Set ratio to 1:1 so height is always equal to width for the preview. */
+	aspect-ratio: 1 / 1;
+}
+
+.woocommerce-coming-soon-entire-site {
+	.wp-container-core-group-is-layout-4 {
+		&.wp-container-core-group-is-layout-4 {
+			justify-content: space-between;
+		}
+	}
+
+	.is-layout-flex {
+		display: flex;
+		flex-wrap: wrap;
+		align-items: center;
+		margin: 0;
+	}
+
+	.wp-block-site-title {
+		p {
+			line-height: normal;
+		}
+
+		a {
+			font-weight: 600;
+			font-size: 20px;
+			font-style: normal;
+			line-height: normal;
+			letter-spacing: -0.4px;
+			color: var(--wp--preset--color--contrast);
+			text-decoration: none;
+		}
+	}
+
+	.wp-block-social-links {
+		gap: 0.5em 18px;
+	}
+
+	.woocommerce-coming-soon-social-login {
+		gap: 48px;
+	}
+
+	.wp-block-loginout {
+		background-color: #000;
+		border-radius: 6px;
+		display: flex;
+		height: 40px;
+		width: 74px;
+		justify-content: center;
+		align-items: center;
+		gap: 10px;
+		box-sizing: border-box;
+
+		a {
+			color: #fff;
+			text-decoration: none;
+			line-height: 17px;
+			font-size: 14px;
+			font-weight: 500;
+		}
+	}
+
+	.wp-block-spacer {
+		margin: 0;
+	}
+
+	.woocommerce-coming-soon-banner-container {
+		padding-inline: min(5.5rem, 8vw);
+		margin: 0;
+		height: 100%;
+		display: flex;
+		flex-direction: column;
+		justify-content: space-between;
+
+		@media (max-width: 660px) {
+			padding-inline: 0;
+		}
+
+		> .wp-block-group__inner-container {
+			height: 100%;
+			display: flex;
+			flex-direction: column;
+			justify-content: space-between;
+		}
+	}
+
+	.woocommerce-coming-soon-powered-by-woo {
+		width: 100%;
+		--wp--preset--spacing--30: 0;
+		--wp--preset--spacing--10: 19px;
+
+		p {
+			font-style: normal;
+			font-weight: 400;
+			line-height: 19.2px;
+			letter-spacing: -0.12px;
+			color: #3c434a;
+			font-size: 12px;
+			font-family: Inter, sans-serif;
+		}
+
+		a {
+			font-family: Inter, sans-serif;
+		}
+	}
+
+	.coming-soon-is-vertically-aligned-center {
+		width: 100%;
+		align-items: stretch;
+	}
+
+	.coming-soon-cover {
+		.wp-block-cover__background {
+			background-color: var(--woocommerce-coming-soon-color) !important;
+		}
+	}
+
+	.woocommerce-coming-soon-header {
+		height: 40px;
+	}
+
+	h1.wp-block-heading.woocommerce-coming-soon-banner {
+		font-size: clamp(27px, 1.74rem + ((1vw - 3px) * 2), 48px);
+		font-weight: 400;
+		line-height: 58px;
+		font-family: Cardo, serif;
+		letter-spacing: normal;
+		text-align: center;
+		font-style: normal;
+		max-width: 820px;
+		color: var(--wp--preset--color--contrast);
+		margin: 0 auto;
+		text-wrap: balance;
+	}
+}
