+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/index.tsx
@@ -10,10 +10,14 @@ import { registerBlockType } from '@wordpress/blocks';
 import Edit from './edit';
 import Save from './save';
 import metadata from './block.json';
+import deprecated from './deprecated';
+import './store-only.scss';
+import './entire-site.scss';
 
 registerBlockType( metadata, {
 	title: __( 'Coming Soon', 'woocommerce' ),
 	edit: Edit,
 	save: Save,
 	apiVersion: 2,
+	deprecated,
 } );
