+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/save.tsx
@@ -3,28 +3,10 @@
  */
 import { useBlockProps, InnerBlocks } from '@wordpress/block-editor';
 
-/**
- * Internal dependencies
- */
-import { generateStyles } from './styles';
-
-export default function Save( { attributes } ) {
-	const { color, storeOnly } = attributes;
-	const blockProps = { ...useBlockProps.save() };
-
-	if ( storeOnly ) {
-		return (
-			<div { ...blockProps }>
-				<InnerBlocks.Content />
-				<style>{ `.woocommerce-breadcrumb {display: none;}` }</style>
-			</div>
-		);
-	}
-
+export default function Save() {
 	return (
-		<div { ...blockProps }>
+		<div { ...useBlockProps.save() }>
 			<InnerBlocks.Content />
-			<style>{ generateStyles( color ) }</style>
 		</div>
 	);
 }
