+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/store-only.scss
@@ -1,0 +1,5 @@
+body:has(.woocommerce-coming-soon-store-only) {
+	.woocommerce-breadcrumb {
+		display: none;
+	}
+}
