+++ plugins/woocommerce-blocks/assets/js/blocks/coming-soon/styles.tsx
@@ -1,3 +1,4 @@
+/** Deprecated styles, please use the styles from the *.scss files. */
 export const generateStyles = ( color = '#bea0f2' ) => {
 	return `
     /* Reset */
@@ -25,10 +26,6 @@ export const generateStyles = ( color = '#bea0f2' ) => {
     body .is-layout-constrained > .alignwide {
         margin: 0 auto;
     }
-	.block-editor-block-preview__content-iframe body {
-		/* Set ratio to 1:1 so height is always equal to width for the preview. */
-		aspect-ratio: 1 / 1;
-	}
     .wp-container-core-group-is-layout-4.wp-container-core-group-is-layout-4 {
         justify-content: space-between;
     }
